Extract CompletedTaskRow in its own component
